% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{coef.zibr_saem}
\alias{coef.zibr_saem}
\title{Coeficientes estimados de un ajuste ZIBR}
\usage{
\method{coef}{zibr_saem}(object, ...)
}
\arguments{
\item{object}{Un objeto \code{zibr_saem}, resultado de \code{\link[=fit_zibr]{fit_zibr()}}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[stats:coef]{stats::coef()}}.}
}
\value{
Vector numerico \code{mu} con los coeficientes de la parte logistica
seguidos de los de la parte beta.
}
\description{
Coeficientes estimados de un ajuste ZIBR
}
