% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{plot.zibr_saem}
\alias{plot.zibr_saem}
\title{Graficar la traza de convergencia de un ajuste ZIBR}
\usage{
\method{plot}{zibr_saem}(x, ...)
}
\arguments{
\item{x}{Un objeto \code{zibr_saem}, resultado de \code{\link[=fit_zibr]{fit_zibr()}}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[=plot]{plot()}}.}
}
\value{
\code{x}, de forma invisible. Se llama por su efecto secundario de
graficar.
}
\description{
Dibuja la traza iteracion-a-iteracion de los parametros del modelo
(\code{x$trace}), util para revisar visualmente la convergencia del algoritmo
SAEM.
}
