% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibbmr.R
\name{print.zibbmr_saem}
\alias{print.zibbmr_saem}
\title{Imprimir un ajuste ZIBBMR}
\usage{
\method{print}{zibbmr_saem}(x, ...)
}
\arguments{
\item{x}{Un objeto \code{zibbmr_saem}, resultado de \code{\link[=fit_zibbmr]{fit_zibbmr()}}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[=print]{print()}}.}
}
\value{
\code{x}, de forma invisible.
}
\description{
Imprimir un ajuste ZIBBMR
}
