% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{print.zibr_saem}
\alias{print.zibr_saem}
\title{Imprimir un ajuste ZIBR}
\usage{
\method{print}{zibr_saem}(x, ...)
}
\arguments{
\item{x}{Un objeto \code{zibr_saem}, resultado de \code{\link[=fit_zibr]{fit_zibr()}}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[=print]{print()}}.}
}
\value{
\code{x}, de forma invisible.
}
\description{
Imprimir un ajuste ZIBR
}
