% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/zibbmr.R, R/zibr.R
\name{se}
\alias{se}
\alias{se.zibbmr_saem}
\alias{se.zibr_saem}
\title{Errores estandar de un ajuste SAEM}
\usage{
se(object, ...)

\method{se}{zibbmr_saem}(object, ...)

\method{se}{zibr_saem}(object, ...)
}
\arguments{
\item{object}{Un objeto ajustado por \code{\link[=fit_zibr]{fit_zibr()}} o \code{\link[=fit_zibbmr]{fit_zibbmr()}} (clases
\code{zibr_saem} o \code{zibbmr_saem}).}

\item{...}{Argumentos adicionales pasados a metodos especificos.}
}
\value{
Un vector numerico con los errores estandar, en el mismo orden que
\code{coef(object)} seguido de la dispersion y las varianzas de efectos
aleatorios.
}
\description{
Generico que calcula los errores estandar de los parametros de un modelo
ajustado, a partir de la matriz de informacion de Fisher estocastica
(\code{fisher_stoch}) almacenada en el objeto de ajuste. Requiere que el modelo
se haya ajustado con \code{compute_fim = TRUE}.
}
\seealso{
\code{\link[=fit_zibr]{fit_zibr()}}, \code{\link[=fit_zibbmr]{fit_zibbmr()}}, \code{\link[stats:vcov]{stats::vcov()}}
}
