% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibbmr.R
\name{vcov.zibbmr_saem}
\alias{vcov.zibbmr_saem}
\title{Matriz de varianza-covarianza de un ajuste ZIBBMR}
\usage{
\method{vcov}{zibbmr_saem}(object, ...)
}
\arguments{
\item{object}{Un objeto \code{zibbmr_saem} ajustado con \code{compute_fim = TRUE}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[stats:vcov]{stats::vcov()}}.}
}
\value{
Una matriz de varianza-covarianza.
}
\description{
Matriz de varianza-covarianza de un ajuste ZIBBMR
}
