% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{vcov.zibr_saem}
\alias{vcov.zibr_saem}
\title{Matriz de varianza-covarianza de un ajuste ZIBR}
\usage{
\method{vcov}{zibr_saem}(object, ...)
}
\arguments{
\item{object}{Un objeto \code{zibr_saem} ajustado con \code{compute_fim = TRUE}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[stats:vcov]{stats::vcov()}}.}
}
\value{
Una matriz de varianza-covarianza.
}
\description{
Calcula la inversa de la matriz de informacion de Fisher estocastica
(\code{fisher_stoch}), estimada durante el ajuste SAEM.
}
