% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibbmr.R
\name{plot.zibbmr_saem}
\alias{plot.zibbmr_saem}
\title{Graficar la traza de convergencia de un ajuste ZIBBMR}
\usage{
\method{plot}{zibbmr_saem}(x, ...)
}
\arguments{
\item{x}{Un objeto \code{zibbmr_saem}, resultado de \code{\link[=fit_zibbmr]{fit_zibbmr()}}.}

\item{...}{No usado, por compatibilidad con el generico \code{\link[=plot]{plot()}}.}
}
\value{
\code{x}, de forma invisible. Se llama por su efecto secundario de
graficar.
}
\description{
Graficar la traza de convergencia de un ajuste ZIBBMR
}
